requirements.txt
# Core dependencies
aiogram==3.13.1
sqlalchemy==2.0.36
alembic==1.14.0
psycopg2-binary==2.9.10
redis==5.2.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
APScheduler==3.10.4
robokassa==1.0.0

# Optional FastAPI
fastapi==0.115.0
uvicorn[standard]==0.32.0

# Admin panel
Django==5.1.2

# Visualizations
numpy==2.1.3
matplotlib==3.9.2


